First commit in the installer branch.
This branch is for improving the initial Foundation installation script.

This commit adds router.php, that acts as a relative .htaccess, so that
we can redirect a user to the installation folder if they don't have
init files.

The usual root .htaccess file is stored in the installation folder by
default, and will be moved by the script.

We also introduce a .gitignore file so that we don't track configuration
files.