Fixed build_image() to give the largest size possible if the requested size does not exist.
If the large size is larger than the image as a whole, the large size
will not be generated. If the user uses a large option in Parsel, the
image would not appear at all, because it would be pointing at an empty
file.

Parsel will now point to the full-size image (which would be the
largest) instead of the empty size.

{ image as large } -> { image as full }