Foundation is now compatible with PHP versions less than 5.3.
Parsel was using variables to pass class names, which is a 5.3 feature.
We've replaced that functionality with call_user_func() which allows us
to do the same thing in earlier versions of PHP5.

This commit also fixes myriad problems and bugs with installation and file access.

The following problems were solved:

• If the user's server uses 'content' in its absolute directory path
(ex. GoDaddy), then Foundation would get confused and not find images.
This is now fixed.

• Parsel wouldn't let the user chain for individual images in the
current project.

• The installation was failing due to a duplicate ID in an
auto-incremented table.

• Small visual bug on install.htm