This commit hotfixes a critical bug that potentially deletes the entire contents of the host server.
On GoDaddy, there was an incident where after deleting test projects,
the entire contents of the server directory disappeared. The cause is,
as of this commit, unknown.

Possible explanations include a malformed or empty project path in the
database that would cause Foundation to recursively remove (rrmdir())
the PUBLIC_PATH global, which is the server's public directory.

The hotfix checks whether the deletion path contains the project slug.
If the project slug is not present in the path, Foundation now displays
a notice telling the user that something is wrong, and that it will not
continue to delete the project files.

It also checks the path in the database and attempts to generate a new
project path in the case that it is empty.

This should be researched further.