This commit fixes myriad errors and problems encountered while working with Frankie Gaw.
• If the user includes any Javascript in the <head> of the document,
Foundation does not include jQuery. We still include cycle.js.

In the future, we should do a more comprehensive check to see if jQuery
specifically is loaded. If it isn't then we can load. If it's an older
version loaded, then we can load the newer version, and then reload the
user's scripts. This is more complicated though.

• Some small changes that automatically lowercases filenames, so that
extension case doesn't get confused in the UI.

• Small change that fixes the save script by allowing the user to save
their email address. (It wasn't saving before)

• !! ffilehandler.js now handles filenames more intelligently, checking for
leading forward-slashes and removing them if necessary before outputting
a filename. There might still be bugs here. !!