Installer actually runs flawlessly now.
There was a problem with when we modified the $path variable for our use
in index.php. It was a bit too soon, and the changes got written into
config.php. That is now fixed and an install has successfully been done
in testing.

We also added some things to the database installer. The site title now
defaults to "My Portfolio", and the projects table has columns for
apparel and fabric for the Fashion Design foundation.